requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
Pillow==11.0.0
gunicorn==23.0.0
sendgrid==6.11.0
